Move destination model receiver methods into same file as the struct
